Add missen files

+++ db/migrate/20160125151440_remove_act_and_actor_localizations.rb
@@ -1,11 +1,13 @@
 class RemoveActAndActorLocalizations < ActiveRecord::Migration
   def change
     if Localization.table_name == 'locations'
-      if Dir.exists?('app/models/localizations')
-        system('rm -rf app/models/localizations')
+      unless Rails.env.test?
+        if Dir.exists?('app/models/localizations')
+          system('rm -rf app/models/localizations')
+        end
+        # Remove config.autoload_paths += Dir[Rails.root.join('app', 'models', 'localizations')] from application.rb
+        system("cp config/application.rb.new config/application.rb") if File.exists?('config/application.rb.new')
       end
-      # Remove config.autoload_paths += Dir[Rails.root.join('app', 'models', 'localizations')] from application.rb
-      system("cp config/application.rb.new config/application.rb") if File.exists?('config/application.rb.new')
       # Drop unused tables
       drop_table :act_localizations
       drop_table :actor_localizations
